requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=0.19.0
anthropic>=0.21.0